# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API Barantin",
        "description": "Data Request",
        "contact": {
            "email": "nandaputraservice@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Api Admin Barantin"
        },
        {
            "url": "http://localhost:8000/api/v2",
            "description": "Api User Barantin Test"
        }
    ],
    "paths": {
...
        "/user/upt": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Ambil Semua User Upt di masing masing user",
                "operationId": "51f44aa2d6e7f471d8106ec5361b23cc",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "master_upt_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Tambahkan Kata \"Bearer\" Sebelum token contoh: (Bearer xx|497zSWISvECINqxxxxxxxxxxxxxxx)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}